% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{sim_mixture}
\alias{sim_mixture}
\title{Simulate a probablistic/continuous change}
\usage{
sim_mixture(n = 50, prob = 0.2, mu1 = 1, sigma1 = 0.25, mu2 = 0, sigma2 = 0.25)
}
\arguments{
\item{n}{Number of simulated values}

\item{prob}{Probability of distribution 1}

\item{mu1}{Mean of distrubtion 1}

\item{sigma1}{sd of distribution 1}

\item{mu2}{mean of distribution2}

\item{sigma2}{sd of distribution 2}
}
\description{
Simulate a probablistic/continuous change
}
