% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{sim_prob_cont}
\alias{sim_prob_cont}
\title{simulate prob/continuous change}
\usage{
sim_prob_cont(
  max_year = 100,
  pop_size = 300,
  phi = 10,
  n = 50,
  mu1 = 1,
  sigma1 = 0.25,
  mu2 = 0,
  sigma2 = 0.25,
  logit_low = -4,
  logit_high = 4
)
}
\arguments{
\item{max_year}{The total number of "years" the change is spread across}

\item{pop_size}{The total number of simulated data points}

\item{phi}{The "precision" of the beta distribution}

\item{n}{Number of simulated values}

\item{mu1}{Mean of distrubtion 1}

\item{sigma1}{sd of distribution 1}

\item{mu2}{mean of distribution2}

\item{sigma2}{sd of distribution 2}
}
\description{
simulate prob/continuous change
}
